refactor(web): restructure routes to use path params instead of search params
Refactor routing structure to use more RESTful path parameters:
- Branches: Use path param for branch selection (/branches/{branch})
- Pull Requests: Split into list and detail routes with path params
- Repository: Add index route for repository overview
- Commits: Rename to index route for consistency

Route structure changes:
- repos.$owner.$repo.tsx → repos.$owner.$repo.index.tsx
- repos.$owner.$repo.branches.tsx → repos.$owner.$repo.branches.$branch.tsx
- repos.$owner.$repo.pull-requests.tsx → split into:
  - repos.$owner.$repo.pull-requests.index.tsx (PR list)
  - repos.$owner.$repo.pull-requests.$pullNumber.index.tsx (PR detail)
- repos.$owner.$repo.commits.$ref.tsx → repos.$owner.$repo.commits.$ref.index.tsx

Benefits:
- More RESTful URL structure
- Better browser history navigation
- Cleaner separation of concerns
- Easier to bookmark and share specific pages

README updated to reflect new route structure.

User request: "search param 대신 path param 사용, Outlet 패턴으로 구조 개선"
