+++ web/src/routes/repos.$owner.$repo.commits.$ref.index.tsx
@@ -17,7 +17,7 @@ import {
 } from "~/components/ui/card";
 import { Skeleton } from "~/components/ui/skeleton";
 
-export const Route = createFileRoute("/repos/$owner/$repo/commits/$ref")({
+export const Route = createFileRoute("/repos/$owner/$repo/commits/$ref/")({
   component: CommitDetailPage,
 });
